--- mobile/screens/__mocks__/api_data.json
@@ -1,35 +1,0 @@
-{
-  "data": {
-    "users": [
-      {
-        "email": "user@teste.com",
-        "first_name": "",
-        "last_name": ""
-      },
-      {
-        "email": "otherUser@teste.com",
-        "first_name": "",
-        "last_name": ""
-      }
-    ],
-    "products": {
-      "id": 1,
-      "fk_vendor": 1,
-      "name": "Brigadeiro",
-      "price": 1.5,
-      "photo": "https://panelinha-sitenovo.s3-sa-east-1.amazonaws.com/receita/958014000000-Brigadeiro.jpg",
-      "description": "nham nham"
-    },
-    "orders": {
-      "id": 1,
-      "date": "2018-11-01T17:59:33.354460Z",
-      "fk_product": 1,
-      "fk_buyer": 1,
-      "buyer_message": "To no RU",
-      "quantity": 1,
-      "total_price": 1.5,
-      "status": 0,
-      "product_name": "Brigadeiro"
-    }
-  }
-}
