+++ mobile/screens/__tests__/Vendas/components/BuyerOrderCard.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import OfferDialog from '../tab_navigator/vendas/components/OfferDialog';
+import BuyerOrderCard from '../../../tab_navigator/vendas/components/BuyerOrderCard';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<OfferDialog />).toJSON();
+  const tree = renderer.create(<BuyerOrderCard />).toJSON();
   expect(tree).toMatchSnapshot();
 });
