+++ mobile/screens/__tests__/Vendas/components/ErrorDialog.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductImage from '../tab_navigator/vendas/components/ProductImage';
+import ErrorDialog from '../../../tab_navigator/vendas/screens/my_products/ErrorDialog';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<ProductImage />).toJSON();
+  const tree = renderer.create(<ErrorDialog />).toJSON();
   expect(tree).toMatchSnapshot();
 });
