+++ mobile/screens/__tests__/Vendas/components/FormPicker.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FormPicker from '../tab_navigator/vendas/components/FormPicker';
+import FormPicker from '../../../tab_navigator/vendas/components/FormPicker';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
