+++ mobile/screens/__tests__/Vendas/components/OfferDialog.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ToogleView from '../tab_navigator/vendas/screens/my_products/ToogleView';
+import OfferDialog from '../../../tab_navigator/vendas/components/OfferDialog';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,7 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const navigation = jest.fn();
-  const tree = renderer.create(<ToogleView navigation={navigation}/>).toJSON();
+  const tree = renderer.create(<OfferDialog />).toJSON();
   expect(tree).toMatchSnapshot();
 });
