+++ mobile/screens/__tests__/Vendas/components/OrderCard.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import BuyerOrderCard from '../tab_navigator/vendas/components/BuyerOrderCard';
+import OrderCard from '../../../tab_navigator/vendas/components/OrderCard';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<BuyerOrderCard />).toJSON();
+  const tree = renderer.create(<OrderCard />).toJSON();
   expect(tree).toMatchSnapshot();
 });
