+++ mobile/screens/__tests__/Vendas/components/ProductCard.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ErrorDialog from '../tab_navigator/vendas/screens/my_products/ErrorDialog';
+import ProductCard from '../../../tab_navigator/vendas/components/ProductCard';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<ErrorDialog />).toJSON();
+  const tree = renderer.create(<ProductCard />).toJSON();
   expect(tree).toMatchSnapshot();
 });
