+++ mobile/screens/__tests__/Vendas/components/ProductImage.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ProductCard from '../tab_navigator/vendas/components/ProductCard';
+import ProductImage from '../../../tab_navigator/vendas/components/ProductImage';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,6 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<ProductCard />).toJSON();
+  const tree = renderer.create(<ProductImage />).toJSON();
   expect(tree).toMatchSnapshot();
 });
