+++ mobile/screens/__tests__/Vendas/components/RedButton.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import GreenButton from '../tab_navigator/vendas/components/GreenButton';
+import RedButton from '../../../tab_navigator/vendas/components/RedButton';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,13 +8,13 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
   it('renders correctly', () => {
-    const tree = renderer.create(<GreenButton />).toJSON();
+    const tree = renderer.create(<RedButton />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('Testing onPress', () => {
     const onPressMock = jest.fn();
-    const wrapper = shallow(<GreenButton onPress = {onPressMock}/>);
+    const wrapper = shallow(<RedButton onPress = {onPressMock}/>);
     const button = wrapper.find('TouchableOpacity').at(0);
     button.simulate('press');
     expect(onPressMock).toBeCalled();
