+++ mobile/screens/__tests__/Vendas/components/ToogleView.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import OrderCard from '../tab_navigator/vendas/components/OrderCard';
+import ToogleView from '../../../tab_navigator/vendas/screens/my_products/ToogleView';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -8,6 +8,7 @@ import renderer from 'react-test-renderer';
 Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
-  const tree = renderer.create(<OrderCard />).toJSON();
+  const navigation = jest.fn();
+  const tree = renderer.create(<ToogleView navigation={navigation}/>).toJSON();
   expect(tree).toMatchSnapshot();
 });
